Update fhirreporter.py
introduced all mapping components.  

self.module_options["all_transcripts"] brings up an error. The picture of it is on discord. So it has been omitted to show that all components are made and valid.

slight editing of codes and systems to better reflect use of Ensembl transcripts instead of Refseq