FHIR-converter MVP
This is push has a working FHIR-converter module.

If a FHIR doc is converted into a OakVar sqlite then back into a FHIR document the given and produced FHIR files are the same. The only differences are the UUIDs. 

summation: FHIR files produced and FHIR files read are the same